pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"langchain>=1.2.15",
	"langchain-community>=0.4.1",
	"langchain-core>=1.2.28",
	"langchain-mcp-adapters>=0.1.9",
	"langchain-openai>=0.3.17",
	"langgraph>=0.2.70",
	"langgraph-checkpoint-redis>=0.4.1",
	"langsmith>=0.1.147",
	"python-dotenv>=1.2.2",
	"ruff>=0.15.11",
]

[tool.setuptools]
packages = ["agent", "tools", "skills"]
